Add 'MP' role to party memberships
The memberships are specifically of them being MP, rather than simply
party members or any other positions in the party (which should have
separate entries). This will also be easier to search/filter on. It
also allows us to simplify the code, as all memberships will have the
same fields.